fix: Dto 수정 및 예외 처리 추가

+++ src/main/java/com/ll/hereispaw/domain/missing/missing/service/MissingService.java
@@ -3,27 +3,20 @@ package com.ll.hereispaw.domain.missing.missing.service;
 import com.ll.hereispaw.domain.missing.Auhtor.entity.Author;
 import com.ll.hereispaw.domain.missing.Auhtor.repository.AuthorRepository;
 import com.ll.hereispaw.domain.missing.missing.dto.request.MissingRequestDTO;
-import com.ll.hereispaw.domain.missing.missing.dto.response.MissingDTO;
+import com.ll.hereispaw.domain.missing.missing.dto.response.MissingResponseDto;
 import com.ll.hereispaw.domain.missing.missing.entity.Missing;
 import com.ll.hereispaw.domain.missing.missing.repository.MissingRepository;
 import com.ll.hereispaw.domain.member.member.entity.Member;
 import com.ll.hereispaw.global.error.ErrorCode;
-import com.ll.hereispaw.global.error.ErrorResponse;
 import com.ll.hereispaw.global.exception.CustomException;
-import com.ll.hereispaw.global.globalDto.GlobalResponse;
-import com.ll.hereispaw.global.webMvc.LoginUser;
-import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.multipart.MultipartFile;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
@@ -33,7 +26,6 @@ import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.UUID;
 
 @Slf4j
@@ -60,13 +52,14 @@ public class MissingService {
     private EntityManager entityManager;
 
     @Transactional
-    public MissingDTO write(Author author, MissingRequestDTO missingRequestDto) {
+    public MissingResponseDto write(Author author, MissingRequestDTO missingRequestDto) {
 
 //        log.debug("author : {}", missingRequestDto.getAuthor());
 
         authorRepository.save(author);
 
         String pathUrl = s3Upload(missingRequestDto.getFile());
+
         Missing missing = missingRepository.save(
                 Missing.builder()
                         .name(missingRequestDto.getName())
@@ -78,7 +71,7 @@ public class MissingService {
                         .gender(missingRequestDto.getGender())
                         .neutered(missingRequestDto.getNeutered())
                         .age(missingRequestDto.getAge())
-//                        .lostDate(missingRequestDto.getLostDate())
+                        .lostDate(missingRequestDto.getLostDate())
                         .etc(missingRequestDto.getEtc())
                         .reward(missingRequestDto.getReward())
                         .missingState(missingRequestDto.getMissingState())
@@ -88,43 +81,47 @@ public class MissingService {
         );
 
 //        s3Upload(missing, file);
-        return new MissingDTO(missing);
+        return new MissingResponseDto(missing);
     }
 
     public Author of(Member member) {
         return entityManager.getReference(Author.class, member.getId());
     }
 
-    public List<MissingDTO> list() {
+    public List<MissingResponseDto> list() {
         List<Missing> missings = missingRepository.findAll();
-        List<MissingDTO> missingDTOS = new ArrayList<>();
+        List<MissingResponseDto> missingResponseDtos = new ArrayList<>();
 
         for (Missing missing : missings) {
 
             Author author = authorRepository.findById(missing.getAuthor().getId()).orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
 
-            missingDTOS.add(
-                    new MissingDTO(missing)
+            missingResponseDtos.add(
+                    new MissingResponseDto(missing)
             );
         }
 
-        return missingDTOS;
+        return missingResponseDtos;
     }
 
-    public MissingDTO findById(Long missingId) {
+    public MissingResponseDto findById(Long missingId) {
         Missing missing = missingRepository.findById(missingId)
                 .orElseThrow(() -> new CustomException(ErrorCode.MISSING_NOT_FOUND));
 
-        return new MissingDTO(missing);
+        return new MissingResponseDto(missing);
     }
 
     @Transactional
-    public MissingDTO update(
+    public MissingResponseDto update(
             Author author,
             MissingRequestDTO missingRequestDTO,
             Long missingId) {
 
         Missing missing = missingRepository.findById(missingId).orElseThrow(() -> new CustomException(ErrorCode.MISSING_NOT_FOUND));
+
+        if (!author.getId().equals(missing.getId())) {
+            throw new CustomException(ErrorCode.METHOD_NOT_ALLOWED);
+        }
 
         String pathUrl = s3Upload(missingRequestDTO.getFile());
 
@@ -145,12 +142,16 @@ public class MissingService {
 
         missingRepository.save(missing);
 
-        return new MissingDTO(missing);
+        return new MissingResponseDto(missing);
     }
 
     @Transactional
     public String delete(Author author, Long missingId) {
         Missing missing = missingRepository.findById(missingId).orElseThrow(() -> new CustomException(ErrorCode.MISSING_NOT_FOUND));
+
+        if (!author.getId().equals(missing.getId())) {
+            throw new CustomException(ErrorCode.METHOD_NOT_ALLOWED);
+        }
 
         s3Delete(missing.getPathUrl());
         missingRepository.delete(missing);
